requirements.txt
boto3
Click
python-dotenv
psycopg2
SQLALchemy
pandas
numpy
scikit-learn
python-dateutil

## DSaPP stuff
git+https://github.com/dssg/triage.git@master
